pkg/operator: Move the op.handleExpirationReport method from operators.go to reports.go.
This felt out of place being in the operators.go with all of the other top-level reporting methods being stored in reports.go.
